+++ src/main/java/com/example/AlexFitness/controller/RequestFitController.java
@@ -8,10 +8,9 @@ import com.example.AlexFitness.service.RequestFitService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @Api("Контроллер для заявок на абонемент и тренера")
 @RestController
@@ -33,13 +32,31 @@ public class RequestFitController {
         requestFitService.createRequest(requestFit);
     }
 
-    /*@GetMapping("/notApproved")
+    @GetMapping("/notApproved")
     public List<RequestFitDTO> findNotApprovedRequests() {
         List<RequestFit> listRequestFit = requestFitService.findNotApprovedRequests();
-        //todo return requestFitMapper.toRequestFitDto();}
+        return requestFitMapper.toRequestFitListDTO(listRequestFit);
+    }
 
+    @ApiOperation("Метод для поиска всех созданных заявок на абонемент и тренера")
+    @GetMapping("/showRequestFit")
+    public List<RequestFit> showRequestFit() {
+        return requestFitService.findNotApprovedRequests();
+    }
 
-    }*/
+    @ApiOperation("Метод для отклонения заявки на абонемент")
+    @PatchMapping("/rejectRequestFit")
+    public void rejectRequestFit(@RequestParam String phoneNumber) {
+        requestFitService.rejectRequestFit(phoneNumber);
+    }
+
+    @ApiOperation("Метод для одобрения заявки на абонемент")
+    @PatchMapping("/approveRequestFit")
+    public void approveRequestFit(@RequestParam String phoneNumber) {
+        requestFitService.approve(phoneNumber);
+    }
 
 
 }
+
+
