+++ src/main/java/com/example/AlexFitness/service/RequestFitService.java
@@ -1,6 +1,7 @@
 package com.example.AlexFitness.service;
 
 
+import com.example.AlexFitness.model.entity.Client;
 import com.example.AlexFitness.model.entity.RequestFit;
 import com.example.AlexFitness.repository.RequestFitRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,10 +13,12 @@ import java.util.List;
 @Service
 public class RequestFitService {
     private final RequestFitRepo requestFitRepo;
+    private final ClientService clientService;
 
     @Autowired
-    public RequestFitService(RequestFitRepo requestFitRepo) {
+    public RequestFitService(RequestFitRepo requestFitRepo, ClientService clientService) {
         this.requestFitRepo = requestFitRepo;
+        this.clientService = clientService;
     }
 
     @Transactional
@@ -33,10 +36,6 @@ public class RequestFitService {
         return requestFitRepo.findAllByIsApprovedNull();
     }
 
-    @Transactional
-    public void approveRequestFit(RequestFit requestFit) {
-        requestFitRepo.save(requestFit);
-    }
 
     @Transactional
     public void rejectRequestFit(String phoneNumber) {
@@ -45,6 +44,20 @@ public class RequestFitService {
             throw new RuntimeException("Заявка с таким номером телефона не найдена");
         }
         requestFit.setApproved(false);
+    }
 
+    @Transactional
+    public void approve(String phoneNumber) {
+        RequestFit requestFit = requestFitRepo.findByPhoneNumber(phoneNumber);
+        if (requestFit == null) {
+            throw new RuntimeException("Заявка с таким номером телефона не найдена");
+        }
+        Client client1 = clientService.findByPhoneNumber(phoneNumber);
+        if (client1 == null) {
+            throw new RuntimeException("Клиент с таким номером телефона не найден");
+        }
+        requestFit.setApproved(true);
+        client1.setCoach(requestFit.getCoachId());
+        client1.setSubscriptionId(requestFit.getSubId());
     }
 }
